# docs/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets": {
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "URGENT"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 160,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTicketsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tickets/images/{name}": {
      "get": {
        "operationId": "TicketsController_image",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tickets/{id}/status": {
      "patch": {
        "operationId": "TicketsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceTicket"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGRA API",
    "description": "Official SIGRA Phase 0 HTTP contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "TicketResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "residentId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "imageName": {
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "type": "string"
          },
          "priority": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "clientRequestId",
          "residentId",
          "description",
          "imageName",
          "status",
          "priority",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedTicketsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "clientRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "maxLength": 2000
          }
        },
        "required": [
          "clientRequestId",
          "description"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "MaintenanceTicket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientRequestId": {
            "type": "string"
          },
          "residentId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientRequestId",
          "residentId",
          "description",
          "imageName",
          "status",
          "priority",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Phase0Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}